+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Core: Added a new property *serverInfo.devMode* to the plugin context that can be used to determine if some packages are in development mode
  * Added a wrapper security provider that adds support for Basic authentication to any other security provider that implements _login()_ properly.
    Useful for end-2-end or load tests and if you want to access some API from an external system.
  * Core: Enable Express "view cache" when no plugin package is in *devMode*,